Include link to strip_heredoc source
Since the method is very small and self contained, it might make sense for the developer to just copy/paste it into the project. I added it as a link so it was less emphasized. I feel like it's better to just use what already exists, but sometimes that doesn't make sense.